refactor: move functional::as into functional/utility.hpp
