+++ include/Simple-Utility/functional/utility.hpp
@@ -1,0 +1,42 @@
+//          Copyright Dominic Koepke 2019 - 2022.
+// Distributed under the Boost Software License, Version 1.0.
+//    (See accompanying file LICENSE_1_0.txt or copy at
+//          https://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef SL_UTILITY_FUNCTIONAL_UTILITY_HPP
+#define SL_UTILITY_FUNCTIONAL_UTILITY_HPP
+
+#pragma once
+
+#include "Simple-Utility/functional/transform.hpp"
+#include "Simple-Utility/concepts/stl_extensions.hpp"
+
+#include <utility>
+
+namespace sl::functional::util
+{
+	/**
+	 * \defgroup GROUP_FUNCTIONAL_UTILITY utility
+	 * \brief Contains functional objects, implementing several utility operations.
+	 * \ingroup GROUP_FUNCTIONAL
+	 * @{
+	 */
+
+	/**
+	 * \brief Functional object which converts the given argument to the target type via static_cast.
+	 * \tparam TTo The target type.
+	 */
+	template <class TTo>
+	inline constexpr transform_fn as{
+		[]<class TFrom>(TFrom&& arg) constexpr noexcept(concepts::nothrow_explicitly_convertible_to<TFrom, TTo>) -> TTo
+		{
+			static_assert(concepts::explicitly_convertible_to<TFrom, TTo>, "Argument is not convertible to target type.");
+
+			return static_cast<TTo>(std::forward<TFrom>(arg));
+		}
+	};
+
+	/** @} */
+}
+
+#endif
