+++ tests/functional/CMakeLists.txt
@@ -6,4 +6,5 @@ target_sources(
 	"transform.cpp"
 	"tuple.cpp"
 	"overloaded.cpp"
+	"utility.cpp"
 )
